#878: Fix issue where gallery media is reversed
