+++ Nos/Views/NoteCard.swift
@@ -188,6 +188,7 @@ struct NoteCard_Previews: PreviewProvider {
                     NoteCard(note: previewData.imageNote, hideOutOfNetwork: false)
                     NoteCard(note: previewData.expiringNote, hideOutOfNetwork: false)
                     NoteCard(note: previewData.verticalImageNote, hideOutOfNetwork: false)
+                    NoteCard(note: previewData.doubleImageNote, hideOutOfNetwork: false)
                     NoteCard(note: previewData.veryWideImageNote, hideOutOfNetwork: false)
                     NoteCard(note: previewData.imageNote, style: .golden, hideOutOfNetwork: false)
                     NoteCard(note: previewData.linkNote, hideOutOfNetwork: false)
